@@ -1,5 +1,5 @@
 class Jobsite {
-  final String title;
-  final String address;
+  final String? title;
+  final String? address;
   Jobsite({this.title, this.address});
 }
